Move TreatmentSelector to its own folder
